requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==22.2.0
Brotli==1.1.0
cffi==1.15.1
charset-normalizer==3.2.0
ciso8601==2.3.0
colorlog==6.7.0
frozenlist==1.4.0
hikari==2.0.0.dev116
hikari-lightbulb==2.3.1
hikari-miru==3.0.2
idna==3.4
multidict==6.0.4
orjson==3.9.7
pycares==4.3.0
pycparser==2.21
python-dotenv==0.21.1
yarl==1.9.2
